rename privacy policy template file
